# Admin-a-main/Admin-a-main/lib/api-spec/openapi.yaml
openapi: "3.0.3"
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetail"
        "404":
          description: User not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /users/{id}/action:
    post:
      operationId: userAction
      tags: [users]
      summary: Perform admin action on user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserActionInput"
      responses:
        "200":
          description: Action performed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/rides:
    get:
      operationId: listUserRides
      tags: [users]
      summary: Get ride history for user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User rides
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RideListResponse"
  /users/{id}/wallet-adjustment:
    post:
      operationId: adjustUserWallet
      tags: [users]
      summary: Add or deduct wallet balance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletAdjustmentInput"
      responses:
        "200":
          description: Wallet updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
...
  /wallets:
    get:
      operationId: listWallets
      tags: [wallets]
      summary: List all wallets
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: frozen
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Wallet list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletListResponse"
  /wallets/{id}:
    get:
      operationId: getWallet
      tags: [wallets]
      summary: Get wallet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Wallet details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallets/{id}/freeze:
    post:
      operationId: freezeWallet
      tags: [wallets]
      summary: Freeze or unfreeze a wallet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletFreezeInput"
      responses:
        "200":
          description: Wallet updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
  /wallets/{id}/transactions:
    get:
      operationId: listWalletTransactions
      tags: [wallets]
      summary: Get wallet transaction history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Wallet transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletTransactionListResponse"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, phone, status, isVerified, totalRides, totalSpending, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        profileImage:
          type: string
        status:
          type: string
          enum: [active, inactive, banned, suspended]
        isVerified:
          type: boolean
        isBlockedFromBooking:
          type: boolean
        totalRides:
          type: integer
        cancelledRides:
          type: integer
        totalSpending:
          type: number
        walletBalance:
          type: number
        fcmToken:
          type: string
        lastActiveAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserDetail:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            recentRides:
              type: array
              items:
                $ref: "#/components/schemas/Ride"
            walletTransactions:
              type: array
              items:
                $ref: "#/components/schemas/WalletTransaction"
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [active, inactive, banned, suspended]
        isVerified:
          type: boolean
        isBlockedFromBooking:
          type: boolean
    UserActionInput:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [ban, unban, suspend, unsuspend, verify, block_booking, unblock_booking, force_logout, reset_account]
        reason:
          type: string
    WalletAdjustmentInput:
      type: object
      required: [amount, type, reason]
      properties:
        amount:
          type: number
        type:
          type: string
          enum: [credit, debit]
        reason:
          type: string
    WalletBalance:
      type: object
      required: [userId, balance]
      properties:
        userId:
          type: string
        balance:
          type: number
    UserListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Ride:
      type: object
      required: [id, status, fare, paymentMethod, paymentStatus, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        driverId:
          type: string
        userName:
          type: string
        userPhone:
          type: string
        driverName:
          type: string
        driverPhone:
          type: string
        pickupAddress:
          type: string
        dropAddress:
          type: string
        pickupLat:
          type: number
        pickupLng:
          type: number
        dropLat:
          type: number
        dropLng:
          type: number
        distanceKm:
          type: number
        fare:
          type: number
        status:
          type: string
          enum: [pending, active, completed, cancelled]
        paymentMethod:
          type: string
          enum: [cash, wallet, online]
        paymentStatus:
          type: string
          enum: [pending, paid, failed, refunded]
        cancelReason:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    RideListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Ride"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Wallet:
      type: object
      required: [id, userId, userName, balance, isFrozen, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: string
        userPhone:
          type: string
        balance:
          type: number
        isFrozen:
          type: boolean
        createdAt:
          type: string
          format: date-time
    WalletListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Wallet"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    WalletFreezeInput:
      type: object
      required: [frozen]
      properties:
        frozen:
          type: boolean
        reason:
          type: string
    WalletTransaction:
      type: object
      required: [id, walletId, type, amount, description, createdAt]
      properties:
        id:
          type: string
        walletId:
          type: string
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: number
        balance:
          type: number
        description:
          type: string
        referenceId:
          type: string
        createdAt:
          type: string
          format: date-time
    WalletTransactionListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/WalletTransaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...